Create Insert() With Recursion
Recursively checks current node data against value to insert. If value
is less than current node data, call recursive function with node.left
passed in. If value is greater than current node data, call recursive
function with node.right passed in. At the end, create a new Node object
with the desired value, with it being the left or right of some node.
There is also an edge case covered where if the value is already equal
to an existing value, return.