Create InOrder()
Create two versions; one with a helper function, and one without that
returns an array in 'inOrder'.
In order is when the left subtree is traversed, then the root, then the
right subtree.
In this case, an empty array is defined, then if the left subtree of a
node exists, then recursively traverse it down to the bottom.
Then return the root,then traverse the right subtree.